Adds bot-parsers as injectionable objects
Handles the botparser equivalent to the client/device parsers. This way
users of the library can add their own botparser that might not be
usefull for the library, but are needed in their own situation.

As there does not seem to be any usecase for adding multple bot-parsers
in this project, there is no separte directory with bot-parsers, nor can
we add a botparser by string.