Adds botparserabstract to implement discarDetails
Re-adds the discardDetails() that accidently got removed in the refactor.
To be able to do this on all added bots, an abstract is created. It could
have been an interface, but all other (client, device) do this with an
abstract, so that pattern is followed

Adds a test that catches the problem in previous commit, and passes now