Adds bot-parsers as injectionable objects (#5712)
* Adds bot-parsers as injectionable objects

Handles the botparser equivalent to the client/device parsers. This way
users of the library can add their own botparser that might not be
usefull for the library, but are needed in their own situation.

As there does not seem to be any usecase for adding multple bot-parsers
in this project, there is no separte directory with bot-parsers, nor can
we add a botparser by string.

* Adds botparserabstract to implement discarDetails

Re-adds the discardDetails() that accidently got removed in the refactor.
To be able to do this on all added bots, an abstract is created. It could
have been an interface, but all other (client, device) do this with an
abstract, so that pattern is followed

Adds a test that catches the problem in previous commit, and passes now