Harmonize version parsing in bot regexes
Bot version parsing in bot regexes is consistently done using `[\d.]+`.

In particular, the following changes are made:

- drop wrapper parentheses, where it is used, as the bot version is not
  exposed by the parser, and there is inconsistent use of grouping
  parentheses in the various regexes;
- `[\d+.]` is assumed to be a typo picked up from the use of `\d+.`
  outside character classes;
- `[\d+\.]` additionally uses a superfluous escape on the dot, which is
  not needed inside character classes;
- plain `\d` or `[0-9]` are replaced with the common expression, as they
  will match the same agent strings, but include more of the version
  string in the match–mainly, this drives consistency;

The main change that should arise here is that the plus character will
no longer be recognised as part of the bot/version match. The tests
don't suggest that this will be an issue. (If needed, the `+` can be
brought back, while keeping consistency)